refactor(framework): Consolidate ABCs and AgentAppCallable into base.py
Agent-Logs-Url: https://github.com/flwrlabs/flower/sessions/b8cb29d5-9efd-40bf-a607-3632ccafe957
