--- framework/py/flwr/agentapp/responses.py
@@ -1,41 +1,0 @@
-# Copyright 2026 Flower Labs GmbH. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ==============================================================================
-"""AgentResponses (abstract base class)."""
-
-
-from __future__ import annotations
-
-from abc import ABC, abstractmethod
-
-from flwr.supercore.typing import JSONObject
-
-
-class AgentResponses(ABC):
-    """Abstract base class for AgentApp model response creation."""
-
-    @abstractmethod
-    def create(self, request: JSONObject) -> JSONObject:
-        """Create a model response.
-
-        Parameters
-        ----------
-        request : JSONObject
-            Open Responses-compatible create request.
-
-        Returns
-        -------
-        response : JSONObject
-            Open Responses-compatible response.
-        """
